requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0
albumentations>=1.3.0
foolbox>=3.3.0
advertorch>=0.2.0
timm>=0.6.0
transformers>=4.20.0
tqdm>=4.62.0
pyyaml>=6.0
click>=8.0.0
jinja2>=3.0.0
plotly>=5.0.0
kaleido>=0.2.1
jupyter>=1.0.0
ipywidgets>=7.6.0 